+++ src/tracer/scene.rs
@@ -8,8 +8,8 @@ use crate::tracer::hit::Hit;
 use crate::tracer::ray::Ray;
 use crate::tracer::texture::Texture;
 use crate::tracer::material::Material;
-use crate::tracer::object::sphere::Sphere;
-use crate::tracer::object::{Object, Plane, Rectangle, Cuboid};
+use crate::tracer::object::{Object, plane::Plane, rectangle::Rectangle};
+use crate::tracer::object::{cuboid::Cuboid, sphere::Sphere};
 
 #[cfg(test)]
 mod scene_tests;
